refactor: job retrieval moved to top-level dir, container logic placed in grading-container/
