+++ grading-vm/grading_container/grading_script/bash_grading_script_command.py
